+++ src/test/java/seedu/reserve/testutil/EditReservationDescriptorBuilder.java
@@ -6,13 +6,13 @@ import java.util.stream.Stream;
 
 import seedu.reserve.logic.commands.EditCommand;
 import seedu.reserve.logic.commands.EditCommand.EditReservationDescriptor;
+import seedu.reserve.model.occasion.Occasion;
 import seedu.reserve.model.reservation.DateTime;
 import seedu.reserve.model.reservation.Diners;
 import seedu.reserve.model.reservation.Email;
 import seedu.reserve.model.reservation.Name;
 import seedu.reserve.model.reservation.Phone;
 import seedu.reserve.model.reservation.Reservation;
-import seedu.reserve.model.tag.Occasion;
 
 /**
  * A utility class to help with building EditReservationDescriptor objects.
